Added copy of LunrIndexer from upstream jekyll-lunar-js-search plugin
MIT licensed, available from:

  https://github.com/slashdotdash/jekyll-lunr-js-search/blob/master/lib/jekyll_lunr_js_search/indexer.rb

The intent is to start from an exact copy, and make the minimal amount
of changes required so that theoretically upstream patches could be
applied. Unfortunately I think the goals of the upstream package are
too broad to consider refactoring their code into a generic package and
submitting patches to them.