Add F-Droid specific indexing to LunrIndexer
The upstream code assumes we want to index pages with a HTML
body. But we want to index app metadata (i.e. name + summary).
The only reason we don't also index the description is because
we need to load the index on each page load, and that might
result in a much larger size.