Remove stemmer which causes problems for app names.
Algorithmic stemmers are great for large bodies of English text. Unfortunately
they seem to be not as useful for small bodies of app  names. These often use
weird spelling and truncation to sound hip and cool, or indeed use words which
are quite unique to this field. As a result there is a bunch of strange edge
cases which came up during testing. This just does away with the stemmer.

The main fear was that this would increase the index size substantially, but
it only means the difference between 746KiB and 778KiB for the F-Droid index.