Removed title from "browse.html".
This is a bit hackey, but the reason is because it is a string which
can't be translated by the website. At this point, it is easier to
remove than to make it translateable. Hopefully it is not a big deal,
because the user still has the menu item "Browse" at the top, which is
underlined indicating that is where they are.