Fix first and last paginator links.
The "first" was incorrectly pointing to "../1" when it should've pointed
to "..". This is because despite there being "/packages/2",
"/packages/3", etc, the first page is just "/packages".

For the exact same reason, the "last" button was broken, but only on
page one. This is because the last page is defined by "../total_pages".
When you are on "/packages/2" this works fine. But when you are on
"/packages" then it breaks.

See https://github.com/sverrirs/jekyll-paginate-v2/issues/28