When searching in the sidebar, correctly redirect to /packages
Prior to this, searching in the sidebar while already on the packages
page wouldn't work because it only changed the has part of the URL and
thus didn't triger a reload. Now it listens for hash changes if
required.

Note that this uses "String.substr" instead of "String.startsWith"
because the later is not supported prior to ES6.