Add repo info to IndexV1, and add repo info tag.
The tag is used via {% fdroid_repo_info %} and outputs
"Repo Name @ updated date". For example: "F-Droid @ 2017-09-11".

Also caches the downloading of repos so that when several pages or
several tests want to create an IndexV1 object for a particular repo,
they needn't worry about multiple downloads. This caching only happens
in the Ruby process, and each new build will start with a clean cache.