Add "what's new" and "suggested_version_name" to App.
Whats new is localized, and suggested_version_name is only set
if there is a relevant suggested version.

Noticed while working on this that suggested version code in the metadata
is a string, whereas each package's version code is an integer, so needed
to cast the suggested code before comparing.

Formatting of "whats new" is a bit off, will need to be fixed when we
fix the formatting of the app descriptions.