Remov inline JS for initializing search, put in external JS file.
The JS was previously inline, so it had access to Jekyll/Liquid
variables such as the unique identifier of the search input wrapper.
Now that it is no longer inline, it doesn't have access to those, so
such variables have been put into `data-*` attributes to be queried via
JS later.