Fix tests which were written against an older GP repo.
Since then two more apps have been published, so changed the expectation
of exactly 10 apps to >= 10 apps when downloading the index fresh from
the internet.

Elsewhere when checking a cached version of index-v1.json, it still just
asserts 10, because that is the index as it was at that point in time.