support Liberapay: field, with migration support from LiberapayID:
Liberapay was originally included using a numeric ID, since they had
not yet finalized the public URLs.  Now it is a username.  So this
logic prefers the username in Liberapay: field, and uses the old
LiberapayID: as a fallback.  LiberapayID: will not override Liberapay:
if it is already set.  This reuses the old database key since it is
stored and processed as a String anyway.