only feature apps in sidebar that have no Anti-Features
Since F-Droid is about free software and privacy, then apps with
Anti-Features like Tracking, NonFreeNet, NonFreeAddon, etc. should not be
promoted extra.  We should highlight the apps that are 100% free software
and respect privacy.  This will still fall back to app that have Anti-
Features if it runs out of other options.